Cluster prototype: handles client-initiated commands (not dequeues)
Details
 - Cluster.cpp: serializes all frames thru cluster (see below)
 - broker/ConnectionManager: Added handler chain in front of Connection::received.
 - sys::Fork and ForkWithMessage - abstractions for forking with posix impl.
 - tests/ForkedBroker.h: test utility to fork a broker process. 
 - broker/SignalHandler: Encapsulated signal handling from qpidd.cpp
 - Various minor logging & error message improvements to aid debugging.

NB: current impl will not scale. It is functional working starting point so we
can start testing & profiling to find the right optimizations.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@674107 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/sys/Socket.h
@@ -47,7 +47,7 @@ public:
     /** Set socket non blocking */
     void setNonblocking() const;
 
-    void connect(const std::string& host, int port) const;
+    void connect(const std::string& host, uint16_t port) const;
 
     void close() const;
 
@@ -67,7 +67,7 @@ public:
      *@param backlog maximum number of pending connections.
      *@return The bound port.
      */
-    int listen(int port = 0, int backlog = 10) const;
+    int listen(uint16_t port = 0, int backlog = 10) const;
     
     /** Returns the "socket name" ie the address bound to 
      * the near end of the socket
